chore(core): move backup flow adapters to `tests.input_flows`
To be re-used in other tests as well.

[no changelog]
