+++ tests/input_flows.py
@@ -14,11 +14,14 @@ from __future__ import annotations
 import time
 from typing import Callable, Generator, Sequence
 
+import pytest
+
 from trezorlib import messages
 from trezorlib.client import Session
 from trezorlib.debuglink import DebugLink, DebugSession, LayoutContent, LayoutType
 from trezorlib.debuglink import TrezorTestContext as Client
 from trezorlib.debuglink import multipage_content
+from trezorlib.exceptions import TrezorFailure
 
 from . import translations as TR
 from .common import (
@@ -34,6 +37,8 @@ from .common import (
 from .input_flows_helpers import BackupFlow, EthereumFlow, PinFlow, RecoveryFlow
 
 B = messages.ButtonRequestType
+
+FlowAdapter = Callable[[Session, Callable[[], BRGeneratorType]], BRGeneratorType]
 
 
 class InputFlowBase:
@@ -3181,3 +3186,47 @@ class InputFlowCancelBrightness(InputFlowBase):
     def input_flow_delizia(self):
         yield
         self.debug.click(self.debug.screen_buttons.menu())
+
+
+# InputFlow adaptors
+
+
+def normal(_session: Session, flow: Callable[[], BRGeneratorType]) -> BRGeneratorType:
+    return flow()
+
+
+def try_to_cancel(skip_cancel: set[str] | None = None) -> FlowAdapter:
+    BACKUP_IN_PROGRESS = messages.Failure(
+        code=messages.FailureType.InProgress,
+        message="Backup in progress",
+    )
+
+    if skip_cancel is None:
+        skip_cancel = set()
+
+    def _try_to_cancel(
+        session: Session, flow: Callable[[], BRGeneratorType]
+    ) -> BRGeneratorType:
+        gen = flow()
+        next(gen)
+        cancels = 0
+        while True:
+            br = yield
+
+            # Don't cancel if the button request appears in `skip_cancel`
+            if br.name not in skip_cancel:
+                # Entering session's context will send an explicit THP ACK after `BACKUP_IN_PROGRESS` is received.
+                with session.client._interact(force_flush=True):
+                    # Try to cancel the backup flow on Core
+                    with pytest.raises(TrezorFailure) as exc_info:
+                        session.call(messages.Cancel(), expect=messages.Failure)
+                    # Following #6483, backup is not cancellable
+                    assert exc_info.value.failure == BACKUP_IN_PROGRESS
+                    cancels += 1
+            try:
+                gen.send(br)
+            except StopIteration:
+                assert cancels > 0
+                return
+
+    return _try_to_cancel
